requirements.txt
rumps
psutil
py2app
pyobjc-framework-Cocoa